Rewind next_index to last entry in receiver log if no entry at prev_log_index.
e.g. the receiver has got

  ... { idx = 10; term = 42 }

at the end of the its log, and receives an Append_entries message that holds

  ... { idx = 100; term = 42 }

A failure message is sent with prev_log_index=10, so that the leader will try
to send entries starting from 11 next.

Note that in case of conflict (not missing entry, as in the above case),
next_index is still rewound by 1 entry per round.